fix: prepare-image page fails to load in production
The Next.js build process uses SWC to minify and optimize code for production. However, the JIMP library appears not to be currently compatible with SWC's minification process, causing a naming conflict that results in a syntax error (`Uncaught SyntaxError: Identifier 'e' has already been declared`).

This issue only occurs in built files. To resolve the issue, SWC minification has been temporarily disabled, forcing Next.js to fall back to using the slower Terser. This allows the code to be compiled correctly and served without errors.

This is a temporary fix and will be reverted in the future.

Link to the related GitHub issue: https://github.com/jimp-dev/jimp/issues/1344